[eas-cli] move queries/mutations for credentials to credentials/ios/api/graphql
