+++ packages/eas-cli/src/credentials/ios/api/graphql/queries/AppleDistributionCertificateQuery.ts
@@ -1,12 +1,12 @@
 import gql from 'graphql-tag';
 
-import { graphqlClient, withErrorHandlingAsync } from '../../client';
+import { graphqlClient, withErrorHandlingAsync } from '../../../../../graphql/client';
 import {
   AppleDistributionCertificate,
   AppleDistributionCertificateFragment,
-} from '../../types/credentials/AppleDistributionCertificate';
-import { AppleTeamFragment } from '../../types/credentials/AppleTeam';
-import { IosDistributionType } from '../../types/credentials/IosAppBuildCredentials';
+} from '../../../../../graphql/types/credentials/AppleDistributionCertificate';
+import { AppleTeamFragment } from '../../../../../graphql/types/credentials/AppleTeam';
+import { IosDistributionType } from '../../../../../graphql/types/credentials/IosAppBuildCredentials';
 
 const AppleDistributionCertificateQuery = {
   async getForAppAsync(
