+++ packages/eas-cli/src/credentials/ios/api/graphql/queries/__mocks__/AppleTeamQuery.ts
@@ -1,4 +1,4 @@
-import { AppleTeam } from '../../../types/credentials/AppleTeam';
+import { AppleTeam } from '../../../../../../graphql/types/credentials/AppleTeam';
 
 const AppleTeamQuery = {
   byAppleTeamIdentifierAsync: jest.fn().mockImplementation(() => {
