+++ packages/eas-cli/src/devices/actions/create/registrationUrlMethod.ts
@@ -4,7 +4,7 @@ import qrcodeTerminal from 'qrcode-terminal';
 import { URL } from 'url';
 
 import { getExpoWebsiteBaseUrl } from '../../../api';
-import { AppleDeviceRegistrationRequestMutation } from '../../../graphql/mutations/credentials/AppleDeviceRegistrationRequestMutation';
+import { AppleDeviceRegistrationRequestMutation } from '../../../credentials/ios/api/graphql/mutations/AppleDeviceRegistrationRequestMutation';
 import { AppleTeam } from '../../../graphql/types/credentials/AppleTeam';
 import log from '../../../log';
